Fixed break vs. continue bug
There was a point in the file loop of the photometry routine that would
cause the loop to break when a bad file was encountered. The loop should
have "continued" with the next good file rather than skipping the rest
of the files in the group with the "break" command.